Feat: StudyBookmark 스키마 및 API 구현
